Implement interfaces for MovieService and MovieRepository
Introduced the IMovieService and IMovieRepository interfaces to the API and refactored MovieService, MovieRepository, and MovieController to use the new interfaces instead of the concrete classes. This change was made to allow for better abstraction, making the code more maintainable and modular, and improving the overall software design. This should also make unit testing easier in the future as we can mock these interfaces.