Added FluentValidation for validating Movie creation
Added FluentValidation and FluentValidation.DependencyInjectionExtensions packages to perform validation on the Movie model during creation. Updated MovieController to incorporate validation logic,where if the movie details do not pass the validation, a warning is logged, and a Bad Request response is returned. Added new MovieValidator that checks if the Id, Title, and YearOfRelease are present and valid. This commit helps in maintaining data integrity and preventing incorrect data entry.